use decimal instead of double

+++ Data/ICredit.cs
@@ -2,12 +2,12 @@ namespace NetCoreWebApi;
 
 public interface ICredit : IAccount
 {
-    public double CreditLimit { get; set; }
-    public double CreditAvailable { get; set; }
-    public double Balance { get; }
+    public decimal CreditLimit { get; set; }
+    public decimal CreditAvailable { get; set; }
+    public decimal Balance { get; }
     public DateTime? LastPaymentDate { get; set; }
     public string? RewardsProgramDetails { get; set; }
     public string? RewardsProgramBalance { get; set; }
-    public double? StatementBalance { get; set; }
+    public decimal? StatementBalance { get; set; }
     public DateTime? StatementDueDate { get; set; }
 }
